# use strconv variants to improve on performance per 'op'